+++ example/aspnet-versioned/build.ps1
@@ -1,0 +1,8 @@
+Set-Location -Path "./TypeSpec"
+
+npm update
+tsp compile .
+
+Set-Location -Path ".."
+
+dotnet build MassiveScale.Petshop.slnx